requirements.txt
# HR Telegram Bot - Dependencies
# Python 3.9+

# Telegram Bot API
pyTelegramBotAPI==4.14.0

# Google Gemini AI
google-genai==0.2.2

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Environment Management
python-dotenv==1.0.0

# Data Generation (for testing)
Faker==22.0.0

# Optional: Better logging
colorlog==6.8.0
